Seperate chapters and courses in admin site
Recently I've noticed that editing notes for a class was becoming
difficult since the pages were so long; in order to combat this, there
are now individual entries for each chapter that can be filtered based
on book. This should make editing easier since you wont have to search
through a long, long web page to find what you're looking for.